Fix counters (#259)
* Fixed usedSenderLimit counter

* Fixed usedCapacity counter

+++ src/main/java/it/pagopa/pn/delayer/model/IncrementUsedSenderLimitDto.java
@@ -1,3 +1,3 @@
 package it.pagopa.pn.delayer.model;
 
-public record IncrementUsedSenderLimitDto(String pk, Long increment, Integer senderLimit) {}
+public record IncrementUsedSenderLimitDto(String pk, Integer increment, Integer senderLimit) {}
